requirements.txt
pynput==1.7.6
pystray==0.17.3
Pillow==8.4.0